Crate leptos_dom

The DOM implementation for leptos.

Re-exports

pub use html::HtmlElement;

Modules

ev

Types for all DOM events.

helpers

A variety of DOM utility functions.

html

Exports types for working with HTML elements.

math

Exports types for working with MathML elements.

ssr

Server-side HTML rendering utilities.

ssr_in_order

Server-side HTML rendering utilities for in-order streaming and async rendering.

svg

Exports types for working with SVG elements.

Macros

debug_warn

Uses println!()-style formatting to log warnings to the console (in the browser) or via eprintln!() (if not in the browser), but only if it’s a debug build.

error

Uses println!()-style formatting to log errors to the console (in the browser) or via eprintln!() (if not in the browser).

log

Uses println!()-style formatting to log something to the console (in the browser) or via println!() (if not in the browser).

warn

Uses println!()-style formatting to log warnings to the console (in the browser) or via eprintln!() (if not in the browser).

Structs

Component

A user-defined leptos component.

ComponentRepr

Custom leptos component.

DynChild

Represents any View that can change over time.

DynChildRepr

The internal representation of the DynChild core-component.

Each

A component for efficiently rendering an iterable.

EachRepr

The internal representation of the Each core-component.

Element

HTML element.

ErrorKey

A unique key for an error that occurs at a particular location in the user interface.

Errors

A struct to hold all the possible errors that could be provided by child Views

Fragment

Represents a group of views.

HydrationCtx

Control and utility methods for hydration.

HydrationKey

A stable identifier within the server-rendering or hydration process.

IntoIter

An owning iterator over all the errors contained in the Errors struct.

Iter

An iterator over all the errors contained in the Errors struct.

NodeRef

Contains a shared reference to a DOM node created while using the view macro to create your UI.

Text

HTML text

Transparent

Wrapper for arbitrary data that can be passed through the view.

Unit

The unit () leptos counterpart.

UnitRepr

The internal representation of the Unit core-component.

Enums

Attribute

Represents the different possible values an attribute node could have.

Class

Represents the different possible values a single class on an element could have, allowing you to do fine-grained updates to single items in Element.classList.

CoreComponent

The core foundational leptos components.

Property

Represents the different possible values an element property could have, allowing you to do fine-grained updates to single fields.

Style

todo docs

View

A leptos view which can be mounted to the DOM.

Traits

CollectView

Collects an iterator or collection into a View.

IntoAttribute

Converts some type into an Attribute.

IntoClass

Converts some type into a Class.

IntoFragment

Trait for converting any iterable into a Fragment.

IntoProperty

Converts some type into a Property.

IntoStyle

Converts some type into a Style.

IntoView

Converts the value into a View.

Functions

console_debug_warn

Log an error to the console (in the browser) or via println!() (if not in the browser), but only in a debug build.

console_error

Log an error to the console (in the browser) or via println!() (if not in the browser).

console_log

Log a string to the console (in the browser) or via println!() (if not in the browser).

console_warn

Log a warning to the console (in the browser) or via println!() (if not in the browser).

create_node_ref

Creates a shared reference to a DOM node created while using the view macro to create your UI.

document

Returns the Document.

is_browser

Returns true if running on the browser (CSR).

is_dev

Returns true if debug_assertions are enabled.

is_release

Returns true if debug_assertions are disabled.

is_server

Returns true if running on the server (SSR).

mount_to

Runs the provided closure and mounts the result to the provided element.

mount_to_body

Runs the provided closure and mounts the result to the <body>.

window

Returns the Window.