Module leptos_dom::ev

Types for all DOM events.

Structs

AnimationEvent

The AnimationEvent class.

BeforeUnloadEvent

The BeforeUnloadEvent class.

CompositionEvent

The CompositionEvent class.

Custom

A custom event.

CustomEvent

The CustomEvent class.

DOMContentLoaded

The DOMContentLoaded event, which receives Event as its argument.

DeviceMotionEvent

The DeviceMotionEvent class.

DeviceOrientationEvent

The DeviceOrientationEvent class.

DragEvent

The DragEvent class.

ErrorEvent

The ErrorEvent class.

Event

The Event class.

FocusEvent

The FocusEvent class.

GamepadEvent

The GamepadEvent class.

HashChangeEvent

The HashChangeEvent class.

InputEvent

The InputEvent class.

KeyboardEvent

The KeyboardEvent class.

MouseEvent

The MouseEvent class.

PageTransitionEvent

The PageTransitionEvent class.

PointerEvent

The PointerEvent class.

PopStateEvent

The PopStateEvent class.

ProgressEvent

The ProgressEvent class.

PromiseRejectionEvent

The PromiseRejectionEvent class.

SecurityPolicyViolationEvent

The SecurityPolicyViolationEvent class.

StorageEvent

The StorageEvent class.

SubmitEvent

The SubmitEvent class.

TouchEvent

The TouchEvent class.

TransitionEvent

The TransitionEvent class.

UiEvent

The UiEvent class.

WheelEvent

The WheelEvent class.

abort

The abort event, which receives UiEvent as its argument.

afterprint

The afterprint event, which receives Event as its argument.

animationcancel

The animationcancel event, which receives AnimationEvent as its argument.

animationend

The animationend event, which receives AnimationEvent as its argument.

animationiteration

The animationiteration event, which receives AnimationEvent as its argument.

animationstart

The animationstart event, which receives AnimationEvent as its argument.

auxclick

The auxclick event, which receives MouseEvent as its argument.

beforeinput

The beforeinput event, which receives InputEvent as its argument.

beforeprint

The beforeprint event, which receives Event as its argument.

beforeunload

The beforeunload event, which receives BeforeUnloadEvent as its argument.

blur

The blur event, which receives FocusEvent as its argument.

canplay

The canplay event, which receives Event as its argument.

canplaythrough

The canplaythrough event, which receives Event as its argument.

change

The change event, which receives Event as its argument.

click

The click event, which receives MouseEvent as its argument.

close

The close event, which receives Event as its argument.

compositionend

The compositionend event, which receives CompositionEvent as its argument.

compositionstart

The compositionstart event, which receives CompositionEvent as its argument.

compositionupdate

The compositionupdate event, which receives CompositionEvent as its argument.

contextmenu

The contextmenu event, which receives MouseEvent as its argument.

copy

The copy event, which receives Event as its argument.

cuechange

The cuechange event, which receives Event as its argument.

cut

The cut event, which receives Event as its argument.

dblclick

The dblclick event, which receives MouseEvent as its argument.

devicemotion

The devicemotion event, which receives DeviceMotionEvent as its argument.

deviceorientation

The deviceorientation event, which receives DeviceOrientationEvent as its argument.

drag

The drag event, which receives DragEvent as its argument.

dragend

The dragend event, which receives DragEvent as its argument.

dragenter

The dragenter event, which receives DragEvent as its argument.

dragleave

The dragleave event, which receives DragEvent as its argument.

dragover

The dragover event, which receives DragEvent as its argument.

dragstart

The dragstart event, which receives DragEvent as its argument.

drop

The drop event, which receives DragEvent as its argument.

durationchange

The durationchange event, which receives Event as its argument.

emptied

The emptied event, which receives Event as its argument.

ended

The ended event, which receives Event as its argument.

error

The error event, which receives ErrorEvent as its argument.

focus

The focus event, which receives FocusEvent as its argument.

focusin

The focusin event, which receives FocusEvent as its argument.

focusout

The focusout event, which receives FocusEvent as its argument.

formdata

The formdata event, which receives Event as its argument.

fullscreenchange

The fullscreenchange event, which receives Event as its argument.

fullscreenerror

The fullscreenerror event, which receives Event as its argument.

gamepadconnected

The gamepadconnected event, which receives GamepadEvent as its argument.

gamepaddisconnected

The gamepaddisconnected event, which receives GamepadEvent as its argument.

gotpointercapture

The gotpointercapture event, which receives PointerEvent as its argument.

hashchange

The hashchange event, which receives HashChangeEvent as its argument.

input

The input event, which receives Event as its argument.

invalid

The invalid event, which receives Event as its argument.

keydown

The keydown event, which receives KeyboardEvent as its argument.

keypress

The keypress event, which receives KeyboardEvent as its argument.

keyup

The keyup event, which receives KeyboardEvent as its argument.

languagechange

The languagechange event, which receives Event as its argument.

load

The load event, which receives Event as its argument.

loadeddata

The loadeddata event, which receives Event as its argument.

loadedmetadata

The loadedmetadata event, which receives Event as its argument.

loadstart

The loadstart event, which receives Event as its argument.

lostpointercapture

The lostpointercapture event, which receives PointerEvent as its argument.

message

The message event, which receives MessageEvent as its argument.

messageerror

The messageerror event, which receives MessageEvent as its argument.

mousedown

The mousedown event, which receives MouseEvent as its argument.

mouseenter

The mouseenter event, which receives MouseEvent as its argument.

mouseleave

The mouseleave event, which receives MouseEvent as its argument.

mousemove

The mousemove event, which receives MouseEvent as its argument.

mouseout

The mouseout event, which receives MouseEvent as its argument.

mouseover

The mouseover event, which receives MouseEvent as its argument.

mouseup

The mouseup event, which receives MouseEvent as its argument.

offline

The offline event, which receives Event as its argument.

online

The online event, which receives Event as its argument.

orientationchange

The orientationchange event, which receives Event as its argument.

pagehide

The pagehide event, which receives PageTransitionEvent as its argument.

pageshow

The pageshow event, which receives PageTransitionEvent as its argument.

paste

The paste event, which receives Event as its argument.

pause

The pause event, which receives Event as its argument.

play

The play event, which receives Event as its argument.

playing

The playing event, which receives Event as its argument.

pointercancel

The pointercancel event, which receives PointerEvent as its argument.

pointerdown

The pointerdown event, which receives PointerEvent as its argument.

pointerenter

The pointerenter event, which receives PointerEvent as its argument.

pointerleave

The pointerleave event, which receives PointerEvent as its argument.

pointerlockchange

The pointerlockchange event, which receives Event as its argument.

pointerlockerror

The pointerlockerror event, which receives Event as its argument.

pointermove

The pointermove event, which receives PointerEvent as its argument.

pointerout

The pointerout event, which receives PointerEvent as its argument.

pointerover

The pointerover event, which receives PointerEvent as its argument.

pointerup

The pointerup event, which receives PointerEvent as its argument.

popstate

The popstate event, which receives PopStateEvent as its argument.

progress

The progress event, which receives ProgressEvent as its argument.

ratechange

The ratechange event, which receives Event as its argument.

readystatechange

The readystatechange event, which receives Event as its argument.

rejectionhandled

The rejectionhandled event, which receives PromiseRejectionEvent as its argument.

reset

The reset event, which receives Event as its argument.

resize

The resize event, which receives UiEvent as its argument.

scroll

The scroll event, which receives Event as its argument.

scrollend

The scrollend event, which receives Event as its argument.

securitypolicyviolation

The securitypolicyviolation event, which receives SecurityPolicyViolationEvent as its argument.

seeked

The seeked event, which receives Event as its argument.

seeking

The seeking event, which receives Event as its argument.

select

The select event, which receives Event as its argument.

selectionchange

The selectionchange event, which receives Event as its argument.

selectstart

The selectstart event, which receives Event as its argument.

slotchange

The slotchange event, which receives Event as its argument.

stalled

The stalled event, which receives Event as its argument.

storage

The storage event, which receives StorageEvent as its argument.

submit

The submit event, which receives SubmitEvent as its argument.

suspend

The suspend event, which receives Event as its argument.

timeupdate

The timeupdate event, which receives Event as its argument.

toggle

The toggle event, which receives Event as its argument.

touchcancel

The touchcancel event, which receives TouchEvent as its argument.

touchend

The touchend event, which receives TouchEvent as its argument.

touchmove

The touchmove event, which receives TouchEvent as its argument.

touchstart

The touchstart event, which receives TouchEvent as its argument.

transitioncancel

The transitioncancel event, which receives TransitionEvent as its argument.

transitionend

The transitionend event, which receives TransitionEvent as its argument.

transitionrun

The transitionrun event, which receives TransitionEvent as its argument.

transitionstart

The transitionstart event, which receives TransitionEvent as its argument.

undelegated

Overrides the EventDescriptor::BUBBLES value to always return false, which forces the event to not be globally delegated.

unhandledrejection

The unhandledrejection event, which receives PromiseRejectionEvent as its argument.

unload

The unload event, which receives Event as its argument.

visibilitychange

The visibilitychange event, which receives Event as its argument.

volumechange

The volumechange event, which receives Event as its argument.

waiting

The waiting event, which receives Event as its argument.

webkitanimationend

The webkitanimationend event, which receives Event as its argument.

webkitanimationiteration

The webkitanimationiteration event, which receives Event as its argument.

webkitanimationstart

The webkitanimationstart event, which receives Event as its argument.

webkittransitionend

The webkittransitionend event, which receives Event as its argument.

wheel

The wheel event, which receives WheelEvent as its argument.

Traits

EventDescriptor

A trait for converting types into web_sys events.