Module leptos_dom::helpers

A variety of DOM utility functions.

Structs

AnimationFrameRequestHandle

Handle that is generated by request_animation_frame_with_handle and can be used to cancel the animation frame request.

IdleCallbackHandle

Handle that is generated by request_idle_callback_with_handle and can be used to cancel the idle callback.

IntervalHandle

Handle that is generated by set_interval and can be used to clear the interval.

TimeoutHandle

Handle that is generated by set_timeout_with_handle and can be used to clear the timeout.

Functions

debounce

“Debounce” a callback function. This will cause it to wait for a period of delay after it is called. If it is called again during that period, it will wait delay before running, and so on. This can be used, for example, to wrap event listeners to prevent them from firing constantly as you type.

event_target

Helper function to extract Event.target from any event.

event_target_checked

Helper function to extract event.target.checked from an event.

event_target_value

Helper function to extract event.target.value from an event.

get_property

Gets the value of a property set on a DOM element.

location

Returns the current window.location.

location_hash

Current window.location.hash without the beginning #.

location_pathname

Current window.location.pathname.

request_animation_frame

Runs the given function between the next repaint using Window.requestAnimationFrame.

request_animation_frame_with_handle

Runs the given function between the next repaint using Window.requestAnimationFrame, returning a cancelable handle.

request_idle_callback

Queues the given function during an idle period using Window.requestIdleCallback.

request_idle_callback_with_handle

Queues the given function during an idle period using Window.requestIdleCallback, returning a cancelable handle.

set_interval

Repeatedly calls the given function, with a delay of the given duration between calls, returning a cancelable handle. See setInterval().

set_interval_with_handle

Repeatedly calls the given function, with a delay of the given duration between calls, returning a cancelable handle. See setInterval().

set_property

Sets a property on a DOM element.

set_timeout

Executes the given function after the given duration of time has passed. setTimeout().

set_timeout_with_handle

Executes the given function after the given duration of time has passed, returning a cancelable handle. setTimeout().

window_event_listener

Creates a window event listener from a typed event.

window_event_listener_untyped

Adds an event listener to the Window, typed as a generic Event.