Struct leptos_dom::html::HtmlElement

pub struct HtmlElement<El: ElementDescriptor> { /* private fields */ }

Represents an HTML element.

Implementations

impl<El: ElementDescriptor + 'static> HtmlElement<El>

pub fn with_view_marker(self, marker: impl Into<String>) -> Self

Adds an optional marker indicating the view macro source.

pub fn into_any(self) -> HtmlElement<AnyElement>

Converts this element into HtmlElement<AnyElement>.

pub fn id(self, id: impl Into<Oco<'static, str>>) -> Self

Adds an id to the element.

pub fn node_ref(self, node_ref: NodeRef<El>) -> Selfwhere Self: Clone,

Binds the element reference to NodeRef.

pub fn on_mount(self, f: impl FnOnce(Self) + 'static) -> Self

Runs the callback when this element has been mounted to the DOM.

Important Note

This method will only ever run at most once. If this element is unmounted and remounted, or moved somewhere else, it will not re-run unless you call this method again.

pub fn is_mounted(&self) -> bool

Checks to see if this element is mounted to the DOM as a child of body.

This method will always return None on non-wasm CSR targets.

pub fn attr( self, name: impl Into<Oco<'static, str>>, attr: impl IntoAttribute ) -> Self

Adds an attribute to this element.

pub fn attrs( self, attrs: impl IntoIterator<Item = (&'static str, Attribute)> ) -> Self

Adds multiple attributes to the element

pub fn class( self, name: impl Into<Oco<'static, str>>, class: impl IntoClass ) -> Self

Adds a class to an element.

Note: In the builder syntax, this will be overwritten by the class attribute if you use .attr("class", /* */). In the view macro, they are automatically re-ordered so that this over-writing does not happen.

Panics

This directly uses the browser’s classList API, which means it will throw a runtime error if you pass more than a single class name. If you want to pass more than one class name at a time, you can use HtmlElement::classes.

pub fn classes(self, classes: impl Into<Oco<'static, str>>) -> Self

Adds a list of classes separated by ASCII whitespace to an element.

pub fn dyn_classes<I, C>(self, classes_signal: impl Fn() -> I + 'static) -> Selfwhere I: IntoIterator<Item = C>, C: Into<Oco<'static, str>>,

Sets the class on the element as the class signal changes.

pub fn style( self, name: impl Into<Oco<'static, str>>, style: impl IntoStyle ) -> Self

Sets a style on an element.

Note: In the builder syntax, this will be overwritten by the style attribute if you use .attr("style", /* */). In the view macro, they are automatically re-ordered so that this over-writing does not happen.

pub fn prop( self, name: impl Into<Oco<'static, str>>, value: impl IntoProperty ) -> Self

Sets a property on an element.

pub fn on<E: EventDescriptor + 'static>( self, event: E, event_handler: impl FnMut(E::EventType) + 'static ) -> Self

Adds an event listener to this element.

pub fn optional_event<E: EventDescriptor + 'static>( self, event: E, event_handler: Option<impl FnMut(E::EventType) + 'static> ) -> Self

Optionally adds an event listener to this element.

Example

#[component]
pub fn Input(
    #[prop(optional)] value: Option<RwSignal<String>>,
) -> impl IntoView {
    view! {  <input/> }
        // only add event if `value` is `Some(signal)`
        .optional_event(
            ev::input,
            value.map(|value| move |ev| value.set(event_target_value(&ev))),
        )
}

pub fn child(self, child: impl IntoView) -> Self

Adds a child to this element.

pub fn inner_html(self, html: impl Into<Oco<'static, str>>) -> Self

Sets the inner HTML of this element from the provided string slice.

Security

Be very careful when using this method. Always remember to sanitize the input to avoid a cross-site scripting (XSS) vulnerability.