Module leptos_dom::ssr

Server-side HTML rendering utilities.

Functions

render_to_stream

Renders a function to a stream of HTML strings.

render_to_stream_with_prefix

Renders a function to a stream of HTML strings. After the view runs, the prefix will run with the same scope. This can be used to generate additional HTML that has access to the same Scope.

render_to_stream_with_prefix_undisposed

Renders a function to a stream of HTML strings and returns the Scope and RuntimeId that were created, so they can be disposed when appropriate. After the view runs, the prefix will run with the same scope. This can be used to generate additional HTML that has access to the same Scope.

render_to_stream_with_prefix_undisposed_with_context

Renders a function to a stream of HTML strings and returns the Scope and RuntimeId that were created, so they can be disposed when appropriate. After the view runs, the prefix will run with the same scope. This can be used to generate additional HTML that has access to the same Scope.

render_to_stream_with_prefix_undisposed_with_context_and_block_replacement

Renders a function to a stream of HTML strings and returns the Scope and RuntimeId that were created, so they can be disposed when appropriate. After the view runs, the prefix will run with the same scope. This can be used to generate additional HTML that has access to the same Scope.

render_to_string

Renders the given function to a static HTML string.